pyproject.toml
[tool.poetry]
name = "tinkerp"
version = "0.1.0"
description = ""
authors = ["Vivek Vaidya"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
snowflake-connector-python = {extras = ["pandas"], version = "^3.14.0"}
cryptography = "^44.0.2"
pyyaml = "^6.0.2"
faker = "^37.1.0"
tenacity = "^9.0.0"
openai = "^1.68.2"
anthropic = "^0.49.0"
psycopg = {extras = ["binary"], version = "^3.2.6"}
numpy = "^2.2.4"
pandas = "^2.2.3"
click = "^8.1.8"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
